+++ CHANGES.md
@@ -1,5 +1,7 @@
 unreleased
 ==========
+  + merlin binary
+    - Add `locate-types` command (#1951)
   + merlin library
     - Fix `merlin_reader` for OpenBSD (#1956)
     - Improve recovery of mutually recursive definitions (#1962, #1963, fixes #1953)